pyproject.toml
[tool.poetry]
name = "fastapi-base-app-by-al1en"
version = "0.1.0"
description = ""
authors = ["zaebaldev <zaebaldev@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standart"], version = "^0.115.12"}
uvicorn = "^0.34.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.40"}
pydantic = {extras = ["email"], version = "^2.11.3"}
alembic = "^1.15.2"
pydantic-settings = "^2.9.1"
asyncpg = "^0.29.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.1"
arq = "^0.25.0"
gunicorn = "^23.0.0"
orjson = "^3.10.16"
uvloop = "^0.21.0"
httpx = "^0.28.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
pre-commit = "^4.2.0"
ruff = "^0.11.6"
pytest-asyncio = "^0.26.0"
faker = "^37.1.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
